+++ MoshiDex/ViewController.m
@@ -133,11 +133,11 @@
 }
 
 #pragma mark UITableView Delegate
-- (int)numberOfSectionsInTableView:(UITableView *)tableView {
+- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
     return 1;
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return moshiRetrievalArray.count;
 }
